début des voter Quack et Duck

+++ src/Controller/QuackController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Quack;
 use App\Form\QuackType;
 use App\Repository\QuackRepository;
+use App\Security\Voter\QuackVoter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -19,6 +20,7 @@ class QuackController extends AbstractController
             'quacks' => $quackRepository->findAll(),
         ]);
     }
+
     #[Route('/quack', name: 'app_quackAdmin_index', methods: ['GET'])]
     public function indexAdmin(QuackRepository $quackRepository): Response
     {
@@ -28,7 +30,7 @@ class QuackController extends AbstractController
     }
 
     #[Route('/quack/new', name: 'app_quack_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, QuackRepository $quackRepository): Response
+    public function new(Request $request, QuackRepository $quackRepository, QuackVoter $quackVoter): Response
     {
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED');
         $user = $this->getUser();
@@ -53,18 +55,24 @@ class QuackController extends AbstractController
     #[Route('/quack/{id}', name: 'app_quack_show', methods: ['GET'])]
     public function show(Quack $quack): Response
     {
-        $this->denyAccessUnlessGranted('IS_AUTHENTICATED');
+        $this->denyAccessUnlessGranted('view', $quack);
         return $this->render('quack/show.html.twig', [
             'quack' => $quack,
         ]);
     }
 
     #[Route('/quack/{id}/edit', name: 'app_quack_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Quack $quack, QuackRepository $quackRepository): Response
+    public function edit(Request $request, Quack $quack, QuackRepository $quackRepository, QuackVoter $quackVoter): Response
     {
-        $this->denyAccessUnlessGranted('IS_AUTHENTICATED');
+        $quackVoter->vote();
+        $this->denyAccessUnlessGranted('edit', $quack);
+        if ($quack->getDuck() !== $this->getUser()) {
+            throw $this->createAccessDeniedException();
+        }
+
         $form = $this->createForm(QuackType::class, $quack);
         $form->handleRequest($request);
+
 
         if ($form->isSubmitted() && $form->isValid()) {
             $quackRepository->save($quack, true);
@@ -82,7 +90,10 @@ class QuackController extends AbstractController
     public function delete(Request $request, Quack $quack, QuackRepository $quackRepository): Response
     {
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED');
-        if ($this->isCsrfTokenValid('delete'.$quack->getId(), $request->request->get('_token'))) {
+        if ($quack->getDuck() !== $this->getUser()) {
+            throw $this->createAccessDeniedException();
+        }
+        if ($this->isCsrfTokenValid('delete' . $quack->getId(), $request->request->get('_token'))) {
             $quackRepository->remove($quack, true);
         }
